LIBITD-2095. Updated Rails from v4.2.6 to 4.2.11.3
Modified the "rails" gem specification in the "Gemfile" to allow
Rails to update to v4.2.11.3, and use Ruby 2.7.4

In addition to updating Rails, this also updated the "nokogiri" gem.

Added "config/initializers/to_time_preserves_timezone.rb" file, as it
is added in new Rails 4.2.11.3 installations (as reported by
https://railsdiff.org/4.2.6/4.2.11.3)

Added a "monkey patch" to "test/test_helper.rb" to enable the tests to
run with Ruby 2.6 and later.

Note: The "rake test" and "rubocop" now produce deprecation warnings
when running.

Updated "Dockerfile" and "Dockerfile.ci" to use "ruby:2.7.4-slim" base
images.

https://issues.umd.edu/browse/LIBITD-2095